==> analysis/14_Day3/analysis/13_fig1/scripts/04_relaxed_piechart.Rmd <==
---
title: "04 Third pie chart with FDR 0.25 events"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
```

Relax the significance threshold to FDR < 0.25 and probability > 0.8.
And see how the pie chart turns out.

```{r}
library(tidyverse)
library(RColorBrewer)
```

## rMATS

```{r}
# The JCEC.txt files of the AS types
path <- '../../06_rMATS/data/rMATS_result/RF14_control'
fns <- file.path(path, list.files(path, 'JCEC.txt'))
```

```{r}
A3SS <- read_tsv(fns[1]) %>%
  filter(FDR < 0.25) %>%
  dplyr::select(chr, longExonStart_0base, shortES, geneSymbol, FDR) %>%
  rename(start = longExonStart_0base, end = shortES, external_gene_name = geneSymbol, significance  = FDR) %>%
  mutate(start = start + 1) %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'AA') %>%
  dplyr::select(-chr, -start,-end)

# A5SS 
A5SS <- read_tsv(fns[2]) %>%
  filter(FDR < 0.25) %>%
  dplyr::select(chr, shortEE, longExonEnd, geneSymbol, FDR) %>%
  rename(start = shortEE, end = longExonEnd, external_gene_name = geneSymbol, significance  = FDR) %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'AD') %>%
  dplyr::select(-chr, -start,-end)

# RI 
RI <- read_tsv(fns[4]) %>%
  filter(FDR < 0.25) %>%
  dplyr::select(chr, upstreamEE, downstreamES, geneSymbol, FDR) %>%
  rename(start = upstreamEE, end = downstreamES, external_gene_name = geneSymbol, significance  = FDR) %>%
   mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'RI') %>%
  dplyr::select(-chr, -start,-end)


# SE
SE <-  read_tsv(fns[5]) %>%
  filter(FDR < 0.25) %>%
  dplyr::select(chr, exonStart_0base, exonEnd, geneSymbol, FDR) %>%
  rename(start = exonStart_0base, end = exonEnd, external_gene_name = geneSymbol, significance  = FDR) %>%
   mutate(start = start + 1) %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'CE') %>%
  dplyr::select(-chr, -start,-end)


# MXE
MXE <- read_tsv(fns[3]) %>%
  filter(FDR < 0.25) %>%
  dplyr::select(chr, upstreamES, downstreamEE, geneSymbol, FDR) %>%
  rename(start = upstreamES, end = downstreamEE, external_gene_name = geneSymbol, significance  = FDR) %>%
   mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'MXE') %>%
  dplyr::select(-chr, -start,-end)


df14 <- bind_rows(A3SS, A5SS, RI, SE, MXE) %>%
  filter(!duplicated(locus))
```

```{r}
# The JCEC.txt files of the AS types
path <- '../../06_rMATS/data/rMATS_result/RF89_control'
fns <- file.path(path, list.files(path, 'JCEC.txt'))
```

```{r}
A3SS <- read_tsv(fns[1]) %>%
  filter(FDR < 0.25) %>%
  dplyr::select(chr, longExonStart_0base, shortES, geneSymbol, FDR) %>%
  rename(start = longExonStart_0base, end = shortES, external_gene_name = geneSymbol, significance  = FDR) %>%
  mutate(start = start + 1) %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'AA') %>%
  dplyr::select(-chr, -start,-end)

# A5SS 
A5SS <- read_tsv(fns[2]) %>%
  filter(FDR < 0.25) %>%
  dplyr::select(chr, shortEE, longExonEnd, geneSymbol, FDR) %>%
  rename(start = shortEE, end = longExonEnd, external_gene_name = geneSymbol, significance  = FDR) %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'AD') %>%
  dplyr::select(-chr, -start,-end)

# RI 
RI <- read_tsv(fns[4]) %>%
  filter(FDR < 0.25) %>%
  dplyr::select(chr, upstreamEE, downstreamES, geneSymbol, FDR) %>%
  rename(start = upstreamEE, end = downstreamES, external_gene_name = geneSymbol, significance  = FDR) %>%
   mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'RI') %>%
  dplyr::select(-chr, -start,-end)


# SE
SE <-  read_tsv(fns[5]) %>%
  filter(FDR < 0.25) %>%
  dplyr::select(chr, exonStart_0base, exonEnd, geneSymbol, FDR) %>%
  rename(start = exonStart_0base, end = exonEnd, external_gene_name = geneSymbol, significance  = FDR) %>%
   mutate(start = start + 1) %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'CE') %>%
  dplyr::select(-chr, -start,-end)


# MXE
MXE <- read_tsv(fns[3]) %>%
  filter(FDR < 0.25) %>%
  dplyr::select(chr, upstreamES, downstreamEE, geneSymbol, FDR) %>%
  rename(start = upstreamES, end = downstreamEE, external_gene_name = geneSymbol, significance  = FDR) %>%
   mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'MXE') %>%
  dplyr::select(-chr, -start,-end)


df89 <- bind_rows(A3SS, A5SS, RI, SE, MXE) %>%
  filter(!duplicated(locus))
```

```{r}
# now find the overlap between the 14 and 89 that is the final results
 
DF <- df14 %>%
  inner_join(df89 %>%
               dplyr::select(-tool, -significance, -external_gene_name,-type), by = 'locus') %>%
  dplyr::select(locus, tool, type, external_gene_name)
```

```{r}
# how many rows are there in the final result of rMATS
nrow(DF)
```

## Whippet 

```{r}
RF14.ctrl <- read_tsv('../../03_ASprofile/data/whippetResult/RF14_vs_cntrl_output.diff') %>% 
  mutate(absDeltaPsi = abs(DeltaPsi)) %>%
  mutate(Gene = str_replace_all(string = Gene, pattern = '\\..+$', replacement = '')) %>%
  mutate(Group = 'RF14.ctrl')  %>%
  filter(absDeltaPsi > 0.1 & Probability > 0.8 & Psi_A > 0.1)  %>%
  mutate(tool = 'Whippet') %>%
  dplyr::select(locus = Coord, tool, type = Type, GeneID = Gene)


RF89.ctrl <- read_tsv('../../03_ASprofile/data/whippetResult/RF89_vs_cntrl_output.diff') %>%
  mutate(absDeltaPsi = abs(DeltaPsi)) %>%
  mutate(Gene = str_replace_all(string = Gene, pattern = '\\..+$', replacement = '')) %>%
  mutate(Group = 'RF89.ctrl')  %>%
  filter(absDeltaPsi > 0.1 & Probability > 0.8 & Psi_A > 0.1) %>%
  dplyr::select(locus = Coord)

overlap_whippet  <-  RF14.ctrl %>%
  inner_join(RF89.ctrl , by = 'locus')


# get the gene name for the table
mart <- useMart("ensembl", dataset = "hsapiens_gene_ensembl")

# the ensembl_gene_id of the genes that you are interested in 
genes <- overlap_whippet$GeneID

# retrieve the feature information of the listed attributes
symbol <- getBM(filters = "ensembl_gene_id",
                attributes = c('ensembl_gene_id','external_gene_name'),
                values = genes, 
                mart = mart)
# add the gene name to the table
Overlap_whippet <- overlap_whippet %>%
  left_join(symbol %>%
              rename(GeneID = ensembl_gene_id), by = 'GeneID') %>%
  dplyr::select(-GeneID)

# how many records are there 
nrow(Overlap_whippet)
```


## IRfinder

```{r}
data_path <- '../../05_IRfinder/data/rawIRfinderResult'
fns <- file.path(data_path, list.files(data_path, pattern = 'txt'))

# format the table like below (generate the locus ID, change the colnames, format the geneID string, separate that one column to three, generate the padj value.)
raw_df_list <- lapply(fns, function(fn){
  res = read_tsv(fn,comment = '#',col_names = T) %>%
    mutate(Locus =paste(Chr,paste(Start, End, sep = '-'), sep = ':' ) ) 
  
  colnames(res) <- colnames(res) %>%
    gsub('-', '_', .)
  
  res = res %>%
    separate('Intron_GeneName/GeneID', into = c('Symbol', 'GeneID', 'Category'), sep = '/') %>%
    # get rid of the . and digits after that in GeneID
    mutate(GeneID = str_replace_all(string = GeneID, pattern = '\\..*$', replacement = '')) %>%
    mutate(padj = p.adjust(p_diff, method = 'BH')) 
})

##################################################################################

# filter the result like below:
# * Remove records marked with "known-exon". 
# * Remove records marked with "MinorIsoform" either in experiment or control. 
# * Adjust the P-value with BH method 
# * Select the records with padj < 0.25

true_df_list <- lapply(raw_df_list, function(df) {
  res = df  %>%
    filter(! Category == 'known-exon') %>%
    filter( A_IRok != 'MinorIsoform' &  B_IRok != 'MinorIsoform' ) %>%
    filter(padj < 0.25) 
  return(res)
} )

false_df_list <- lapply(raw_df_list, function(df) {
  res = df  %>%
    filter(! Category == 'known-exon') %>%
    filter( A_IRok != 'MinorIsoform' &  B_IRok != 'MinorIsoform' ) %>%
    filter(padj >= 0.25) 
  return(res)
} )

##################################################################################

# TRUE: sig in both
# FALSE: sig in neither but overlapped in both

true_locus <- intersect(true_df_list[[1]]$Locus, true_df_list[[2]]$Locus)
false_locus <- intersect(false_df_list[[1]]$Locus, false_df_list[[2]]$Locus)

# subset the df to include only the true and false locus

subset_df_list <- lapply(raw_df_list, function(df){
  res = df %>%
    filter(Locus %in% true_locus) 
})

overlap_IRfinder <- subset_df_list[[1]] %>%
  mutate(tool = 'IRfinder',
         type = "RI") %>%
  dplyr::select(locus = Locus, tool, type, external_gene_name = Symbol)



# correct the IRfinder start position
Overlap_IRfinder <- overlap_IRfinder %>%
  separate(locus, into = c('chr','coord'), sep = ':') %>%
  separate(coord, into = c('start', 'end'), sep = '-', convert = T) %>%
  mutate(start = start + 1) %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':'))  %>%
  dplyr::select(locus, tool, type, external_gene_name)

# how many records are there
nrow(Overlap_IRfinder)
```

 
```{r}
# merging the whippet and IRfinder result together and remove duplicates if any
whi <- Overlap_whippet
IR <- Overlap_IRfinder

dup <- intersect(whi$locus, IR$locus)

two <- bind_rows(whi,
                 IR %>%
                   filter(! locus %in% dup)) 
```

## Merging the three

```{r}
dup <- intersect(two$locus, DF$locus)


final <- bind_rows(two,
                 DF %>%
                   filter(! locus %in% dup)) 
```

```{r}
hex <- c(brewer.pal(n = 8, name = "Dark2"), '#00468B')
# draw a pie chart of that
colorPal <- data_frame(type = c('AL','AD','TS','MXE','RI','AF','CE','AA','TE'), 
  color = hex) 

res <- final %>%
          group_by(type) %>%
          summarise(count = n())%>%
          mutate(perc = paste0(round((count / sum(count)) * 100), '%') ) 
  
res <- colorPal %>%
  left_join(res, by = 'type') %>%
          mutate(lbl = paste(type, perc , sep = ' '))

png('../figs/FDR025 Splicing type pie chart in SIG events.png',pointsize = 8, res = 300, width = 4, height = 4, units = 'in')
pie(res$count,labels = res$lbl, col= res$color, main = 'FDR025 Splicing type pie chart in SIG events')
dev.off()

write_csv(res, '../output/type_dist_FDR025.csv')
```

```{r out.width=400}
knitr::include_graphics('../figs/FDR025 Splicing type pie chart in SIG events.pdf')
```

```{r }
# a side by side comparison on the raw count of the type distribution
dist005 <- read_csv('../output/type_dist_FDR005.csv') %>%
  mutate(Group = 'FDR005') %>%
  dplyr::select(type, count, Group)
dist025 <- read_csv('../output/type_dist_FDR025.csv') %>%
  mutate(Group = 'FDR025') %>%
  dplyr::select(type, count, Group)

dist <- bind_rows(dist005, dist025) %>%
  mutate(Group =  factor(Group))

dist %>%
  mutate(type = factor(type, levels = c('RI','CE','TE','TS','AA','AD','AF','AL','MXE'))) %>%
  ggplot(aes(x = type, y = count, fill = Group)) +
  geom_bar(stat="identity", position=position_dodge()) +
  scale_fill_manual(values =  c('#00468B', '#EC0000'))  +
  labs(title = 'Type distribution in threshold FDR 0.05 and 0.25') +
  ggsave('../figs/type dist of the two.jpg', width=10, height = 6, dpi = 300)
 
```

```{r}
# unique gene IDs of the sig 025 locus
unique_names <- final %>%
  filter(!duplicated(external_gene_name)) %>%
  pull(external_gene_name)

mart <- useMart("ensembl", dataset = "hsapiens_gene_ensembl")

# the ensembl_gene_id of the genes that you are interested in 
genes <- unique_names

# retrieve the feature information of the listed attributes
symbol <- getBM(filters = "external_gene_name",
                attributes = c('ensembl_gene_id','external_gene_name'),
                values = genes, 
                mart = mart)

ret <- symbol %>%
  filter(!duplicated(external_gene_name)) %>%
  pull(ensembl_gene_id)

write.table(ret, '../data/union_sig_genes_from_3_analysis_relaxed.txt', col.names = F, row.names = F, quote = F)
```
