==> analysis/14_Day3/analysis/13_fig1/scripts/09_assemble_figs.Rmd <==
---
title: "09 figs"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc_depth: 3
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
library(tidyverse)
library(kableExtra)
```

# Left: Day 3. Right: Day 5.

## Filtered significant data

```{r out.width='50%' }
# the sig 05 pie chart
knitr::include_graphics(c('../figs/FDR005 Splicing type pie chart in SIG events.png','../../../../13_fig1/figs/FDR005 Splicing type pie chart in SIG events.png' ))

```

```{r out.width='50%'}
knitr::include_graphics(c('../figs/AS type of the FDR < 0.05 locus.jpg','../../../../13_fig1/figs/AS type of the FDR < 0.05 locus.jpg'))
```

The GSEA of the significant FDR < 0.05 genes.
```{r out.width='50%'}
# the gsea heatmap
knitr::include_graphics(c('../output/Metascape/day3_sig_05_gene_gsea/Enrichment_heatmap/HeatmapSelectedGO.png','../../../../13_fig1/output/Metascape/sig05_2288_unique_gene/Enrichment_heatmap/HeatmapSelectedGO.png'))
```

```{r out.width='50%'}
# the 005 and 025 bar plot
knitr::include_graphics(c('../figs/type dist of the two.jpg','../../../../13_fig1/figs/type dist of the two.jpg'))
```


## Unfiltered data

```{r out.width='50%'}
# the dist bar plot
knitr::include_graphics(c('../figs/AS type distribution in KD and control samples grouped bar plot.jpg','../../../../13_fig1/figs/AS type distribution in KD and control samples grouped bar plot.jpg'))
```

## Overlap between the two

```{r out.width='40%'}
knitr::include_graphics(c('../figs/Day 3 VS Day 5 Comparison on RI.png','../figs/Day 3 VS Day 5 Comparison on CE.png'))
knitr::include_graphics(c('../figs/Day 3 VS Day 5 Comparison on TE.png','../figs/Day 3 VS Day 5 Comparison on TS.png'))
knitr::include_graphics(c('../figs/Day 3 VS Day 5 Comparison on AA.png','../figs/Day 3 VS Day 5 Comparison on AD.png'))
knitr::include_graphics(c('../figs/Day 3 VS Day 5 Comparison on MXE.png','../figs/Day 3 VS Day 5 Comparison on AF.png'))
knitr::include_graphics('../figs/Day 3 VS Day 5 Comparison on AL.png')
```

## Blue intron

```{r out.width='50%'}
# the pie chart
knitr::include_graphics(c('../figs/haveExpression VS unfiltered sig genes detk.png','../../../../13_fig1/figs/haveExpression VS unfiltered sig genes detk.png'))
```

```{r out.width='50%'}
knitr::include_graphics(c('../figs/log10(mean TPM) of have expression only and intersection genes.jpg','../../../../13_fig1/figs/log10(mean TPM) of have expression only and intersection genes.jpg'))
```

**For the genes that have at least 3 introns and have TPM > 50 in all three in BLUE**

```{r out.width='50%'}
knitr::include_graphics(c('../output/Metascape/day3_blue_intron_gt3_tpm50/Enrichment_heatmap/HeatmapSelectedGO.png','../../../../13_fig1/output/Metascape/d5_blue_intron3_tpm50/Enrichment_heatmap/HeatmapSelectedGO.png'))
 
```

**For the genes that have at least 3 introns and have TPM > 50 in all three in PINK**

```{r out.width='50%'}
knitr::include_graphics(c('../output/Metascape/day3_pink_intron_3_50/Enrichment_heatmap/HeatmapSelectedGO.png','../../../../13_fig1/output/Metascape/d5_pink_intron_3_50/Enrichment_heatmap/HeatmapSelectedGO.png'))

```


For the genes that have at least 1 intron and have TPM > 20 in all three in BLUE

```{r out.width='50%'}
knitr::include_graphics(c('../output/Metascape/day3_blue_intron_gt1_tpm20/Enrichment_heatmap/HeatmapSelectedGO.png','../../../../13_fig1/output/Metascape/d5_blue_intron1_tpm20/Enrichment_heatmap/HeatmapSelectedGO.png'))
 
```


```{r}
# what are the translation genes in d3-3inton-tpm50

pull_hits_genes <- function(df, desp){
  ret = df %>%
    filter(Description == desp) %>%
    pull(Hits) %>%
    strsplit( '\\|') %>% 
    unlist
  
  return(ret)
}

library(biomaRt)
mart <- useMart("ensembl", dataset = "hsapiens_gene_ensembl")
genes <- pull_hits_genes(read_csv('../output/Metascape/day3_blue_intron_gt3_tpm50/Enrichment_GO/GO_AllLists.csv'),'Translation')
# retrieve the feature information of the listed attributes
symbol <- getBM(filters = "external_gene_name",
                attributes = c('external_gene_name', 'description'),
                values = genes, 
                mart = mart)

symbol %>%
  kable( caption = 'What are the genes in *Translation* in day 3 with 3 introns and TPM at least 50') %>%
  kable_styling(full_width = F) 
```
