==> analysis/14_Day3/analysis/13_fig1/scripts/18_what_is_wrong_with_rMATS_PSI.Rmd <==
---
title: "Check what is going on with the problematic closer to 0.5 PSI thing"
author: "Anqi Dai"
date: "5/13/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
library(tidyverse)
```

# rMATS intersection

```{r}
# now do the above for the rMATS result
# The JCEC.txt files of the AS types
path <- '../../06_rMATS/data/rMATS_result/RF14_control'
fns <- file.path(path, list.files(path, 'JCEC.txt'))
# A3SS 
A3SS <- read_tsv(fns[1]) %>%
  dplyr::select(chr, longExonStart_0base, shortES, geneSymbol, FDR, IncLevel1, IncLevel2) %>%
  filter(FDR < 0.05) %>% 
  rename(start = longExonStart_0base, end = shortES, external_gene_name = geneSymbol, Psi_A = IncLevel1, Psi_B = IncLevel2) %>%
  separate(Psi_A, into = c('Psi_A1','Psi_A2','Psi_A3'), sep = ',', convert = T) %>%
  separate(Psi_B, into = c('Psi_B1','Psi_B2','Psi_B3'), sep = ',', convert = T) %>%
  rowwise() %>%
  mutate(Psi_A = mean(c(Psi_A1, Psi_A2, Psi_A3), na.rm = T),
         Psi_B = mean(c(Psi_B1, Psi_B2, Psi_B3), na.rm = T))  %>%
  mutate(start = start + 1) %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'AA') %>%
  dplyr::select(external_gene_name:type)
# A5SS 
A5SS <- read_tsv(fns[2]) %>%
  dplyr::select(chr, shortEE, longExonEnd, geneSymbol, FDR, IncLevel1, IncLevel2) %>%
  filter(FDR < 0.05) %>% 
  rename(start = shortEE, end = longExonEnd, external_gene_name = geneSymbol,  Psi_A = IncLevel1, Psi_B = IncLevel2) %>%
  separate(Psi_A, into = c('Psi_A1','Psi_A2','Psi_A3'), sep = ',', convert = T) %>%
  separate(Psi_B, into = c('Psi_B1','Psi_B2','Psi_B3'), sep = ',', convert = T) %>%
  rowwise() %>%
  mutate(Psi_A = mean(c(Psi_A1, Psi_A2, Psi_A3), na.rm = T),
         Psi_B = mean(c(Psi_B1, Psi_B2, Psi_B3), na.rm = T))  %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'AD') %>%
  dplyr::select(external_gene_name:type)
  
# RI 
RI <- read_tsv(fns[4]) %>%
  dplyr::select(chr, upstreamEE, downstreamES, geneSymbol, FDR, IncLevel1, IncLevel2) %>%
  filter(FDR < 0.05) %>% 
  rename(start = upstreamEE, end = downstreamES, external_gene_name = geneSymbol, Psi_A = IncLevel1, Psi_B = IncLevel2) %>%
  separate(Psi_A, into = c('Psi_A1','Psi_A2','Psi_A3'), sep = ',', convert = T) %>%
  separate(Psi_B, into = c('Psi_B1','Psi_B2','Psi_B3'), sep = ',', convert = T) %>%
  rowwise() %>%
  mutate(Psi_A = mean(c(Psi_A1, Psi_A2, Psi_A3), na.rm = T),
         Psi_B = mean(c(Psi_B1, Psi_B2, Psi_B3), na.rm = T))  %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'RI') %>%
  dplyr::select(external_gene_name:type)
# SE
SE <-  read_tsv(fns[5]) %>%
  dplyr::select(chr, exonStart_0base, exonEnd, geneSymbol, FDR, IncLevel1, IncLevel2) %>%
  filter(FDR < 0.05) %>% 
  rename(start = exonStart_0base, end = exonEnd, external_gene_name = geneSymbol, Psi_A = IncLevel1, Psi_B = IncLevel2) %>%
  separate(Psi_A, into = c('Psi_A1','Psi_A2','Psi_A3'), sep = ',', convert = T) %>%
  separate(Psi_B, into = c('Psi_B1','Psi_B2','Psi_B3'), sep = ',', convert = T) %>%
  rowwise() %>%
  mutate(Psi_A = mean(c(Psi_A1, Psi_A2, Psi_A3), na.rm = T),
         Psi_B = mean(c(Psi_B1, Psi_B2, Psi_B3), na.rm = T))  %>%
  mutate(start = start + 1) %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'CE') %>%
  dplyr::select(external_gene_name:type)
# MXE
MXE <- read_tsv(fns[3]) %>%
  dplyr::select(chr, upstreamES, downstreamEE, geneSymbol, FDR, IncLevel1, IncLevel2) %>%
  filter(FDR < 0.05) %>% 
  rename(start = upstreamES, end = downstreamEE, external_gene_name = geneSymbol, Psi_A = IncLevel1, Psi_B = IncLevel2) %>%
  separate(Psi_A, into = c('Psi_A1','Psi_A2','Psi_A3'), sep = ',', convert = T) %>%
  separate(Psi_B, into = c('Psi_B1','Psi_B2','Psi_B3'), sep = ',', convert = T) %>%
  rowwise() %>%
  mutate(Psi_A = mean(c(Psi_A1, Psi_A2, Psi_A3), na.rm = T),
         Psi_B = mean(c(Psi_B1, Psi_B2, Psi_B3), na.rm = T))  %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'MXE') %>%
  dplyr::select(external_gene_name:type)
```

```{r}
df14 <- bind_rows(A3SS, A5SS, RI, SE, MXE) %>%
  filter(!duplicated(locus)) %>%
  dplyr::select(locus, type, Psi_A, Psi_B, external_gene_name)


```

```{r}
# The JCEC.txt files of the AS types
path <- '../../06_rMATS/data/rMATS_result/RF89_control'
fns <- file.path(path, list.files(path, 'JCEC.txt'))
# A3SS 
A3SS <- read_tsv(fns[1]) %>%
  dplyr::select(chr, longExonStart_0base, shortES, geneSymbol, FDR, IncLevel1, IncLevel2) %>%
  filter(FDR < 0.05) %>% 
  rename(start = longExonStart_0base, end = shortES, external_gene_name = geneSymbol, Psi_A = IncLevel1, Psi_B = IncLevel2) %>%
  separate(Psi_A, into = c('Psi_A1','Psi_A2','Psi_A3'), sep = ',', convert = T) %>%
  separate(Psi_B, into = c('Psi_B1','Psi_B2','Psi_B3'), sep = ',', convert = T) %>%
  rowwise() %>%
  mutate(Psi_A = mean(c(Psi_A1, Psi_A2, Psi_A3), na.rm = T),
         Psi_B = mean(c(Psi_B1, Psi_B2, Psi_B3), na.rm = T))  %>%
  mutate(start = start + 1) %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'AA') %>%
  dplyr::select(external_gene_name:type)
# A5SS 
A5SS <- read_tsv(fns[2]) %>%
  dplyr::select(chr, shortEE, longExonEnd, geneSymbol, FDR, IncLevel1, IncLevel2) %>%
  filter(FDR < 0.05) %>% 
  rename(start = shortEE, end = longExonEnd, external_gene_name = geneSymbol,  Psi_A = IncLevel1, Psi_B = IncLevel2) %>%
  separate(Psi_A, into = c('Psi_A1','Psi_A2','Psi_A3'), sep = ',', convert = T) %>%
  separate(Psi_B, into = c('Psi_B1','Psi_B2','Psi_B3'), sep = ',', convert = T) %>%
  rowwise() %>%
  mutate(Psi_A = mean(c(Psi_A1, Psi_A2, Psi_A3), na.rm = T),
         Psi_B = mean(c(Psi_B1, Psi_B2, Psi_B3), na.rm = T))  %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'AD') %>%
  dplyr::select(external_gene_name:type)
  
# RI 
RI <- read_tsv(fns[4]) %>%
  dplyr::select(chr, upstreamEE, downstreamES, geneSymbol, FDR, IncLevel1, IncLevel2) %>%
  filter(FDR < 0.05) %>% 
  rename(start = upstreamEE, end = downstreamES, external_gene_name = geneSymbol, Psi_A = IncLevel1, Psi_B = IncLevel2) %>%
  separate(Psi_A, into = c('Psi_A1','Psi_A2','Psi_A3'), sep = ',', convert = T) %>%
  separate(Psi_B, into = c('Psi_B1','Psi_B2','Psi_B3'), sep = ',', convert = T) %>%
  rowwise() %>%
  mutate(Psi_A = mean(c(Psi_A1, Psi_A2, Psi_A3), na.rm = T),
         Psi_B = mean(c(Psi_B1, Psi_B2, Psi_B3), na.rm = T))  %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'RI') %>%
  dplyr::select(external_gene_name:type)
# SE
SE <-  read_tsv(fns[5]) %>%
  dplyr::select(chr, exonStart_0base, exonEnd, geneSymbol, FDR, IncLevel1, IncLevel2) %>%
  filter(FDR < 0.05) %>% 
  rename(start = exonStart_0base, end = exonEnd, external_gene_name = geneSymbol, Psi_A = IncLevel1, Psi_B = IncLevel2) %>%
  separate(Psi_A, into = c('Psi_A1','Psi_A2','Psi_A3'), sep = ',', convert = T) %>%
  separate(Psi_B, into = c('Psi_B1','Psi_B2','Psi_B3'), sep = ',', convert = T) %>%
  rowwise() %>%
  mutate(Psi_A = mean(c(Psi_A1, Psi_A2, Psi_A3), na.rm = T),
         Psi_B = mean(c(Psi_B1, Psi_B2, Psi_B3), na.rm = T))  %>%
  mutate(start = start + 1) %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'CE') %>%
  dplyr::select(external_gene_name:type)
# MXE
MXE <- read_tsv(fns[3]) %>%
  dplyr::select(chr, upstreamES, downstreamEE, geneSymbol, FDR, IncLevel1, IncLevel2) %>%
  filter(FDR < 0.05) %>% 
  rename(start = upstreamES, end = downstreamEE, external_gene_name = geneSymbol, Psi_A = IncLevel1, Psi_B = IncLevel2) %>%
  separate(Psi_A, into = c('Psi_A1','Psi_A2','Psi_A3'), sep = ',', convert = T) %>%
  separate(Psi_B, into = c('Psi_B1','Psi_B2','Psi_B3'), sep = ',', convert = T) %>%
  rowwise() %>%
  mutate(Psi_A = mean(c(Psi_A1, Psi_A2, Psi_A3), na.rm = T),
         Psi_B = mean(c(Psi_B1, Psi_B2, Psi_B3), na.rm = T))  %>%
  mutate(locus = paste(chr, paste(start,end, sep='-'),sep=':')) %>%
  mutate(tool = 'rMATS',
         type = 'MXE') %>%
  dplyr::select(external_gene_name:type)
```

```{r}
df89 <- bind_rows(A3SS, A5SS, RI, SE, MXE) %>%
  filter(!duplicated(locus))  %>%
  dplyr::select(locus, type_89 = type, Psi_A_89 = Psi_A, Psi_B_89 = Psi_B)

rmats <- df14 %>%
  inner_join(df89, by  = 'locus') %>%
  filter(type == type_89) %>%
  mutate(dis14_A = abs(Psi_A - 0.5),
         dis89_A = abs(Psi_A_89 - 0.5),
         dis14_B = abs(Psi_B - 0.5),
         dis89_B = abs(Psi_B_89 - 0.5)) %>%
  mutate(Psi_A =  ifelse(dis14_A < dis89_A,Psi_A, Psi_A_89 ),
         Psi_B =  ifelse(dis14_B < dis89_B,Psi_B, Psi_B_89 )) %>%
  mutate(deltaPSI = Psi_A - Psi_B)


rmats %>%
  filter( external_gene_name == 'MKNK2')  
```

```{r}
# "keep the PSI_A that is closer to 0.5"

ret <- rMATS_ol  %>%
  mutate(dis14 = abs(Psi_14 - 0.5),
         dis89 = abs(Psi_89 - 0.5)) %>%
  mutate(Psi_A =  ifelse(dis14 < dis89, Psi_14, Psi_89 ),
         deltaPSI = Psi_A - Psi_B) 

# find the gene ENSG00000099875: MKNK2
target <- ret %>%
  filter( ensembl_gene_id == 'ENSG00000099875')

```

Conclusion: my analysis is correct, what is wrong is the data itself from rMATS
